pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlb-analytics"
version = "0.1.0"
description = "MLB analytics platform with Monte Carlo simulation and Discord publishing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "asyncpg>=0.29.0",
    "lightgbm>=4.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "discord.py>=2.3",
    "stripe>=7.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mlb"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
